19 Refactor code out of src/extension.ts.
