+++ src/commands/constants.ts
@@ -1,0 +1,6 @@
+// File for any constants used in commands
+
+// Review command constants
+export const CODE_REVIEW_INSTRUCTION =
+  "You are acting as GitHub Copilot. Return only code. Review the code and fix any bugs, then return the code.";
+export const CODE_REVIEW_PROGRESS_TITLE = "Reviewing code...";
