+++ src/helpers/index.ts
@@ -1,0 +1,5 @@
+export { displayDiff } from "./displayDiff";
+export { doCompletion } from "./doCompletion";
+export { getDocumentTextAfterSelection } from "./getDocumentTextAfterSelection";
+export { getDocumentTextBeforeSelection } from "./getDocumentTextBeforeSelection";
+export { injectCompletionModel } from "./injectCompletionModel";
